Fase D do PRD, D3: ícones comprimidos, 985 KB de volta e a Fase D fechada
Os 1805 ícones de Pokémon (incluindo as gens 6 a 9) comprimiram para smol
(1420 KB para 434,9 KB) com o carregamento redesenhado sem um byte de RAM
estática nova: CreateMonIcon aloca os dois quadros na OBJ VRAM e
descomprime direto para lá (animar virou trocar tileNum, sumiu o DMA por
quadro), e os consumidores de CPU usam buffer de heap que morre na própria
chamada. A troca de tipo para const u32* fez o compilador varrer os
consumidores; sete adaptados. Três DestroySprite de ícone que já eram bug
antes (devolviam faixa errada de tiles) viraram FreeAndDestroyMonIconSprite,
e o fechador ainda pegou um quarto defeito real: o trainer card enfileirava
DMA adiado apontando para heap já liberado, agora cópia imediata com o
endereço derivado do template do bg.

Prova visual no emulador: o ícone do Pikachu desenhado no party menu,
cores conferidas na PLTT OBJ, idêntico à ROM pré-D3. Suíte 336/337
(T11.3 pulado), T11 completo 3/3, SAVE COMPATIVEL.

Consertos de fechamento da fase: o Pikachu do salto de capítulo nascia
DESMAIADO com maxHP 0 (faltava o CalculateMonStats que o CreateMon deste
fork deixa para o chamador; bug do seletor, não do D3) e o T85.3 foi
endurecido pinando a flag de derrotado do camper DARIEN na precondição,
aposentando o flake documentado.

Placar da Fase D inteira: D1 484 KB + D2 101 KB + D3 985 KB = ~1,57 MB
recuperados sem perder um byte de conteúdo. ROM: de 97,69% para 92,90%
(31.171.408 B, ~2,38 MB livres). A meta da fase (~1,3 MB) foi superada e
a obra de Galar cabe com folga, inclusive com DLC.

+++ include/pokemon.h
@@ -445,13 +445,13 @@ struct SpeciesInfo /*0xC4*/
     const u32 *backPic;
     const u16 *palette;
     const u16 *shinyPalette;
-    const u8 *iconSprite;
+    const u32 *iconSprite;
 #if P_GENDER_DIFFERENCES
     const u32 *frontPicFemale;
     const u32 *backPicFemale;
     const u16 *paletteFemale;
     const u16 *shinyPaletteFemale;
-    const u8 *iconSpriteFemale;
+    const u32 *iconSpriteFemale;
 #endif //P_GENDER_DIFFERENCES
 #if P_FOOTPRINTS
     const u8 *footprint;
@@ -529,7 +529,7 @@ struct SpeciesInfo /*0xC4*/
 
 struct EggData
 {
-    const u8 *eggIcon;
+    const u32 *eggIcon;
     const u32 *eggSprite;
     const u16 *eggPalette;
     const u32 *eggHatchGfx;
